feat: Complete stock predictor application (Phases 1-5)
Full-stack stock analysis and prediction platform with:
- FastAPI backend with SQLAlchemy ORM and PostgreSQL
- Next.js 14 frontend with React Query and Recharts
- Celery task queue for async data fetching
- Technical indicators (RSI, MACD, SMA, EMA)
- Backtesting engine with equity curve visualization
- Watchlist and alert system
- Yahoo Finance and NewsAPI integrations

242 backend tests passing, all frontend routes compile.

+++ backend/app/models/price.py
@@ -12,7 +12,7 @@ class StockPrice(Base):
 
     __tablename__ = "stock_prices"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     stock_id = Column(Integer, ForeignKey("stocks.id", ondelete="CASCADE"), nullable=False, index=True)
     date = Column(Date, nullable=False, index=True)
